requirements.txt
semantic-kernel
azure-storage-blob
python-dotenv
pandas
tabulate
